+++ packages/cli/lib/commands/start.ts
@@ -21,16 +21,15 @@ const execSyncNpmStart = (port: number, options: ExecSyncOptions): void => {
 
 const command: StartCommandType = {
 	command: "start",
-	describe: "starts the project",
+	describe: "Starts the project",
 	templateManager: null,
 	builder: (yargs) => {
 		return yargs
 			.option("port", {
 				alias: "p",
-				describe: "serve app port",
+				describe: "Port to serve the app on",
 				type: "number"
-			})
-			.usage(""); // do not show any usage instructions before the commands
+			});
 	},
 	async handler(argv: ArgumentsCamelCase<PositionalArgs>) {
 		GoogleAnalytics.post({