+++ packages/igx-templates/igx-ts/projects/side-nav-auth/files/src/app/authentication/providers/facebook-provider.ts
@@ -51,6 +51,6 @@ export class FacebookProvider implements AuthProvider {
   }
 
   public logout() {
-    FB.logout((response) => { });
+    FB.logout();
   }
 }